Move rendering logic from example into blade-render
