+++ blade-render/Cargo.toml
@@ -10,7 +10,9 @@ repository = "https://github.com/kvark/blade"
 [lib]
 
 [dependencies]
-bg = { package = "blade-graphics", version = "0.2", path = "../blade-graphics"}
+blade = { package = "blade-graphics", version = "0.2", path = "../blade-graphics"}
+blade-macros = { version = "0.2", path = "../blade-macros"}
+bytemuck = { workspace = true }
 gltf = { version = "1.1", features = ["names", "utils"] }
 glam = { workspace = true }
 log = { workspace = true }
