+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub use blade_graphics::*;
 pub use blade_macros::*;
-//TODO: only re-export blade-render?
 //pub use blade_render::*;
